+++ include/DeprecatedGUI/CCheckbox.h
@@ -84,7 +84,7 @@ public:
 
 
 	// Process a message sent
-	DWORD SendMessage(int iMsg, DWORD Param1, DWORD Param2) {
+	uintptr_t SendMessage(int iMsg, uintptr_t Param1, uintptr_t Param2) {
 
 				switch(iMsg) {
 					case CKM_SETCHECK:
@@ -97,8 +97,8 @@ public:
 
 				return 0;
 			}
-	DWORD SendMessage(int iMsg, const std::string& sStr, DWORD Param) { return 0; }
-	DWORD SendMessage(int iMsg, std::string *sStr, DWORD Param)  { return 0; }
+	uintptr_t SendMessage(int iMsg, const std::string& sStr, uintptr_t Param) { return 0; }
+	uintptr_t SendMessage(int iMsg, std::string *sStr, uintptr_t Param)  { return 0; }
 
 
 	// Draw the title button
