+++ include/DeprecatedGUI/CGuiSkinnedLayout.h
@@ -94,9 +94,9 @@ public:
 	int		KeyUp(UnicodeChar c, int keysym, const ModifiersState& modstate);
 
 	// Other functions are empty, because GUI layout don't do any actions itself
-	DWORD	SendMessage(int iMsg, DWORD Param1, DWORD Param2) { return 0; };
-	DWORD	SendMessage(int iMsg, const std::string& sStr, DWORD Param) { return 0; };
-	DWORD	SendMessage(int iMsg, std::string *sStr, DWORD Param) { return 0; };
+	uintptr_t	SendMessage(int iMsg, uintptr_t Param1, uintptr_t Param2) { return 0; };
+	uintptr_t	SendMessage(int iMsg, const std::string& sStr, uintptr_t Param) { return 0; };
+	uintptr_t	SendMessage(int iMsg, std::string *sStr, uintptr_t Param) { return 0; };
 
 	void	LoadStyle() { };
