+++ src/client/DeprecatedGUI/Menu_Player.cpp
@@ -197,7 +197,7 @@ void Menu_PlayerInitialize()
 	cViewPlayers.SendMessage(vp_Name,TXM_SETMAX,20,0);
 	cViewPlayers.SendMessage(vp_Name, TXM_SETFLAGS, TXF_NOUNICODE, 0); // Disable unicode for backward compatibility
 
-	cViewPlayers.SendMessage(vp_Players,		LVM_SETOLDSTYLE, (DWORD)0, 0);
+	cViewPlayers.SendMessage(vp_Players,		LVM_SETOLDSTYLE, (uintptr_t)0, 0);
 
     // Hide the AI stuff until 'Computer' type is selected
     cViewPlayers.getWidget(vp_AIDiffLbl)->setEnabled(false);
@@ -321,7 +321,7 @@ void Menu_Player_ViewPlayerInit()
     Menu_Player_FillSkinCombo( (CCombobox *)cViewPlayers.getWidget(vp_PlySkin) );
 
     // Set the name of the first item in the list
-    int sel = (int)cViewPlayers.SendMessage( vp_Players, LVM_GETCURINDEX,(DWORD)0,0);
+    int sel = (int)cViewPlayers.SendMessage( vp_Players, LVM_GETCURINDEX,(uintptr_t)0,0);
 	SmartPointer<profile_t> p = FindProfile(sel);
     if(p.get()) {
         cViewPlayers.SendMessage( vp_Name,    TXS_SETTEXT, p->sName,0);
@@ -404,8 +404,8 @@ void Menu_Player_NewPlayer(int mouse)
                     cNewPlayer.SendMessage(np_PlySkin, CBS_GETCURSINDEX, &skin, 0);
 
 					// Add the profile
-                    WormType* type = ( (int)cNewPlayer.SendMessage(np_Type,CBM_GETCURINDEX,(DWORD)0,0) == PRF_HUMAN->toInt() ) ? PRF_HUMAN : PRF_COMPUTER;
-                    int level = (int)cNewPlayer.SendMessage(np_AIDiff,SLM_GETVALUE,(DWORD)0,0);
+                    WormType* type = ( (int)cNewPlayer.SendMessage(np_Type,CBM_GETCURINDEX,(uintptr_t)0,0) == PRF_HUMAN->toInt() ) ? PRF_HUMAN : PRF_COMPUTER;
+                    int level = (int)cNewPlayer.SendMessage(np_AIDiff,SLM_GETVALUE,(uintptr_t)0,0);
 
 					AddProfile(name, skin, "", "",r, g, b, type->toInt(), level);
 
@@ -423,7 +423,7 @@ void Menu_Player_NewPlayer(int mouse)
             case np_Type:
                 if(ev->iEventMsg == CMB_CHANGED) {
 
-                    int type = (int)cNewPlayer.SendMessage(np_Type,CBM_GETCURINDEX,(DWORD)0,0);
+                    int type = (int)cNewPlayer.SendMessage(np_Type,CBM_GETCURINDEX,(uintptr_t)0,0);
 
                     // Hide the AI stuff if it is a human type of player
                     cNewPlayer.getWidget(np_AIDiffLbl)->setEnabled(type == PRF_COMPUTER->toInt());
@@ -492,10 +492,10 @@ void Menu_Player_NewPlayer(int mouse)
 
 
     // Draw the difficulty level
-    int type = (int)cNewPlayer.SendMessage(np_Type,CBM_GETCURINDEX,(DWORD)0,0);
+    int type = (int)cNewPlayer.SendMessage(np_Type,CBM_GETCURINDEX,(uintptr_t)0,0);
     if( type == PRF_COMPUTER->toInt() ) {
         static const std::string difflevels[] = {"Easy", "Medium", "Hard", "Xtreme"};
-        int level = (int)cNewPlayer.SendMessage(np_AIDiff,SLM_GETVALUE,(DWORD)0,0);
+        int level = (int)cNewPlayer.SendMessage(np_AIDiff,SLM_GETVALUE,(uintptr_t)0,0);
         tLX->cFont.Draw(VideoPostProcessor::videoSurface().get(), 250,363,tLX->clNormalLabel,difflevels[level]);
 
     }
@@ -592,7 +592,7 @@ void Menu_Player_ViewPlayers(int mouse)
 						}
 
                         // Update the details
-                        int sel = (int)cViewPlayers.SendMessage(vp_Players,LVM_GETCURINDEX,(DWORD)0,0);
+                        int sel = (int)cViewPlayers.SendMessage(vp_Players,LVM_GETCURINDEX,(uintptr_t)0,0);
 	                    p = FindProfile(sel);
                         if(p.get()) {
                             cViewPlayers.SendMessage( vp_Name,		TXS_SETTEXT,    p->sName,0);
@@ -627,7 +627,7 @@ void Menu_Player_ViewPlayers(int mouse)
             // Apply button
             case vp_Apply:
                 if( ev->iEventMsg == BTN_CLICKED ) {
-                    int sel = (int)cViewPlayers.SendMessage(vp_Players, LVM_GETCURINDEX, (DWORD)0,0);
+                    int sel = (int)cViewPlayers.SendMessage(vp_Players, LVM_GETCURINDEX, (uintptr_t)0,0);
 	                SmartPointer<profile_t> p = FindProfile(sel);
 	                if(p.get()) {
 
@@ -648,11 +648,11 @@ void Menu_Player_ViewPlayers(int mouse)
                         p->sName = name;
                         cViewPlayers.SendMessage(vp_Name, TXS_SETTEXT, p->sName, 0);
 
-                        p->R = (Uint8)cViewPlayers.SendMessage(vp_Red,SLM_GETVALUE,(DWORD)0,0);
-                        p->G = (Uint8)cViewPlayers.SendMessage(vp_Green,SLM_GETVALUE,(DWORD)0,0);
-                        p->B = (Uint8)cViewPlayers.SendMessage(vp_Blue,SLM_GETVALUE,(DWORD)0,0);
+                        p->R = (Uint8)cViewPlayers.SendMessage(vp_Red,SLM_GETVALUE,(uintptr_t)0,0);
+                        p->G = (Uint8)cViewPlayers.SendMessage(vp_Green,SLM_GETVALUE,(uintptr_t)0,0);
+                        p->B = (Uint8)cViewPlayers.SendMessage(vp_Blue,SLM_GETVALUE,(uintptr_t)0,0);
                         p->iType = ((CCombobox *)cViewPlayers.getWidget(vp_Type))->getSelectedIndex();
-                        p->nDifficulty = (int)cViewPlayers.SendMessage(vp_AIDiff, SLM_GETVALUE,(DWORD)0,0);
+                        p->nDifficulty = (int)cViewPlayers.SendMessage(vp_AIDiff, SLM_GETVALUE,(uintptr_t)0,0);
 
 						// Reload the graphics
 						std::string buf;
@@ -662,7 +662,7 @@ void Menu_Player_ViewPlayers(int mouse)
 						p->cSkin.ColorizeDefault();
 
                         // Update the item
-                        lv_item_t *it = (lv_item_t *)cViewPlayers.SendMessage(vp_Players, LVM_GETCURITEM, (DWORD)0,0); // TODO: 64bit unsafe (pointer cast)
+                        lv_item_t *it = (lv_item_t *)cViewPlayers.SendMessage(vp_Players, LVM_GETCURITEM, (uintptr_t)0,0);
                         if(it) {
                             if(it->tSubitems) {
 								it->tSubitems->bmpImage = p->cSkin.getPreview();
@@ -691,7 +691,7 @@ void Menu_Player_ViewPlayers(int mouse)
             // Player listbox
             case vp_Players:
                 if( ev->iEventMsg == LV_CHANGED ) {
-                    int sel = (int)cViewPlayers.SendMessage(vp_Players,LVM_GETCURINDEX,(DWORD)0,0);
+                    int sel = (int)cViewPlayers.SendMessage(vp_Players,LVM_GETCURINDEX,(uintptr_t)0,0);
 	                SmartPointer<profile_t> p = FindProfile(sel);
                     if(p.get()) {
                         cViewPlayers.SendMessage( vp_Name,		TXS_SETTEXT,    p->sName,0);
@@ -716,7 +716,7 @@ void Menu_Player_ViewPlayers(int mouse)
             case vp_Type:
                 if( ev->iEventMsg == CMB_CHANGED ) {
 
-                    int type = (int)cViewPlayers.SendMessage(vp_Type,CBM_GETCURINDEX,(DWORD)0,0);
+                    int type = (int)cViewPlayers.SendMessage(vp_Type,CBM_GETCURINDEX,(uintptr_t)0,0);
 
                     // Hide the AI stuff if it is a human type of player
                     cViewPlayers.getWidget(vp_AIDiffLbl)->setEnabled(type == PRF_COMPUTER->toInt());
@@ -794,10 +794,10 @@ void Menu_Player_ViewPlayers(int mouse)
 	}
 
     // Draw the difficulty level
-    int type = (int)cViewPlayers.SendMessage(vp_Type,CBM_GETCURINDEX,(DWORD)0,0);
+    int type = (int)cViewPlayers.SendMessage(vp_Type,CBM_GETCURINDEX,(uintptr_t)0,0);
     if( type == PRF_COMPUTER->toInt() ) {
         static const std::string difflevels[] = {"Easy", "Medium", "Hard", "Xtreme"};
-        int level = (int)(DWORD)CLAMP(cViewPlayers.SendMessage(vp_AIDiff,SLM_GETVALUE,(DWORD)0,0), (DWORD)0, (DWORD)3);
+        int level = (int)(uintptr_t)CLAMP(cViewPlayers.SendMessage(vp_AIDiff,SLM_GETVALUE,(uintptr_t)0,0), (uintptr_t)0, (uintptr_t)3);
         tLX->cFont.Draw(VideoPostProcessor::videoSurface().get(), 530,313,tLX->clNormalLabel, difflevels[level]);
     }
